1. Moved All essential functions to separate files. 2. Updated licence to Apache version 2 from MIT. 3. Improved documentation.
