+++ flower-core/src/commonMain/kotlin/com/hadiyarajesh/flower_core/NetworkResource.kt
@@ -16,7 +16,6 @@
 
 package com.hadiyarajesh.flower_core
 
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 /**
@@ -25,6 +24,7 @@ import kotlinx.coroutines.flow.flow
  * @author Rajesh Hadiya
  * @param makeNetworkRequest - A function to make network request
  * @param onNetworkRequestFailed - An action to perform when a network request fails
+ * @return [REMOTE] type
  */
 inline fun <REMOTE> networkResource(
     crossinline makeNetworkRequest: suspend () -> ApiResponse<REMOTE>,
@@ -52,45 +52,6 @@ inline fun <REMOTE> networkResource(
 
         is ApiEmptyResponse -> {
             emit(Resource.emptySuccess())
-        }
-    }
-}
-
-/**
- * Make a network request and emit the response. Additionally, takes an action to perform
- * if a network request fails.
- * @author Rajesh Hadiya
- * @param makeNetworkRequest - A function to make network request
- * @param onNetworkRequestFailed - An action to perform when a network request fails
- */
-inline fun <REMOTE> networkResourceFlow(
-    crossinline makeNetworkRequest: () -> Flow<ApiResponse<REMOTE>>,
-    crossinline onNetworkRequestFailed: (errorBody: String?, statusCode: Int) -> Unit = { _: String?, _: Int -> }
-) = flow<Resource<REMOTE>> {
-    emit(Resource.loading(null))
-
-    makeNetworkRequest().collect { apiResponse ->
-        when (apiResponse) {
-            is ApiSuccessResponse -> {
-                apiResponse.body?.let {
-                    emit(Resource.success(data = it))
-                }
-            }
-
-            is ApiErrorResponse -> {
-                onNetworkRequestFailed(apiResponse.errorMessage, apiResponse.statusCode)
-                emit(
-                    Resource.error(
-                        msg = apiResponse.errorMessage,
-                        statusCode = apiResponse.statusCode,
-                        null
-                    )
-                )
-            }
-
-            is ApiEmptyResponse -> {
-                emit(Resource.emptySuccess())
-            }
         }
     }
 }
